+++ lib/flutter_lua.dart
@@ -3,14 +3,16 @@
 /// Lua interpreter for Flutter.
 library flutter_lua;
 
-export 'src/errors.dart';
+export 'src/errors.dart' show LuaError;
 export 'src/thread.dart' show LuaThread;
 
 import 'dart:async' show Future;
 import 'dart:io' show File;
 
-import 'package:flutter/services.dart' show MethodChannel;
+import 'package:flutter/services.dart' show MethodChannel, PlatformException;
 import 'package:meta/meta.dart' show experimental;
+
+import 'src/errors.dart' show LuaError;
 
 const MethodChannel _channel = const MethodChannel('flutter_lua');
 
@@ -22,12 +24,20 @@ abstract class Lua {
 
   @experimental
   static Future<void> doFile(final File path) async {
-    final absolutePath = path.isAbsolute ? path : path.absolute;
-    return await _channel.invokeMethod('doFile', absolutePath.toString());
+    try {
+      final absolutePath = path.isAbsolute ? path : path.absolute;
+      return await _channel.invokeMethod('doFile', absolutePath.toString());
+    } on PlatformException catch (error) {
+      throw LuaError.from(error);
+    }
   }
 
   @experimental
   static Future<void> doString(final String code) async {
-    return await _channel.invokeMethod('doString', code);
+    try {
+      return await _channel.invokeMethod('doString', code);
+    } on PlatformException catch (error) {
+      throw LuaError.from(error);
+    }
   }
 }