+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.2.0] - 2019-03-10
 ### Added
+- `LuaError` exception class
+### Changed
+- Implemented `PlatformException` wrapping
 
 ## [0.1.0] - 2019-02-27
 ### Added